+++ CHANGELOG.rst
@@ -19,6 +19,8 @@ Version 0.0.4, current
 - add ``is_valid()`` and ``handle_invalid`` to ``yews.datasets.BaseDatasets``
 - add ``yews.datasets.utils`` to handle urls and bz2 files.
 - add ``yews.datasets.wenchuan``
+- add ``memory limit`` to ``yews.datasets`` to determine when to load the
+  entire dataset into memory.
 
 Version 0.0.3, 2019-04-18
 -------------------------